+++ .gitignore
@@ -160,4 +160,6 @@ cython_debug/
 #.idea/
 
 # Files creates by process
-Files/
+backups/
+devices/
+reports/
