+++ setup.cfg
@@ -1,5 +1,5 @@
 [flake8]
-ignore = E999,E712,
+ignore = E999,
 max-line-length = 120
 max-complexity = 10
