+++ tests/2.simulator/test_simulator.py
@@ -1,0 +1,3 @@
+
+def test_modulo_simulator(app):
+    assert app.simulator._start_simulator() is True
