+++ tests/3.dashboard/test_process.py
@@ -1,0 +1,3 @@
+def test_modulo_process(app):
+
+    assert app.report._start_process().message.get("state") is True
