--- tests/dashboard/test_process.py
@@ -1,3 +1,0 @@
-def test_modulo_process(app):
-
-    assert app.report._start_process().message.get("state") == True
