--- tests/simulator/test_simulator.py
@@ -1,3 +1,0 @@
-
-def test_modulo_simulator(app):
-    assert app.simulator._start_simulator() == True
